refactor claim details container
